do refactoring
